Extract subprocess runner
